third_party/WebKit/LayoutTests/external/wpt/feature-policy/experimental-features/unsized-image.tentative.https.sub-expected.txt

This is a testharness.js-based test.
FAIL Test image with attribute undefined=undefined and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.jpg assert_equals: width: expected 300 but got 64
FAIL Test image with attribute undefined=undefined and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.png assert_equals: width: expected 300 but got 32
FAIL Test image with attribute undefined=undefined and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.svg assert_equals: width: expected 300 but got 769
FAIL Test video with attribute undefined=undefined and attribute undefined=undefined assert_equals: width: expected 300 but got 320
FAIL Test image with attribute width=500 and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.jpg assert_equals: height: expected 150 but got 500
FAIL Test image with attribute width=500 and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.png assert_equals: height: expected 150 but got 500
FAIL Test image with attribute width=500 and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.svg assert_equals: height: expected 150 but got 375
FAIL Test video with attribute width=500 and attribute undefined=undefined assert_equals: height: expected 150 but got 375
FAIL Test image with attribute height=800 and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.jpg assert_equals: width: expected 300 but got 800
FAIL Test image with attribute height=800 and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.png assert_equals: width: expected 300 but got 800
FAIL Test image with attribute height=800 and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.svg assert_equals: width: expected 300 but got 1067
FAIL Test video with attribute height=800 and attribute undefined=undefined assert_equals: width: expected 300 but got 1066.65625
FAIL Test image with attribute style=width:500px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.jpg assert_equals: height: expected 150 but got 500
FAIL Test image with attribute style=width:500px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.png assert_equals: height: expected 150 but got 500
FAIL Test image with attribute style=width:500px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.svg assert_equals: height: expected 150 but got 375
FAIL Test video with attribute style=width:500px; and attribute undefined=undefined assert_equals: height: expected 150 but got 375
FAIL Test image with attribute style=height:800px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.jpg assert_equals: width: expected 300 but got 800
FAIL Test image with attribute style=height:800px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.png assert_equals: width: expected 300 but got 800
FAIL Test image with attribute style=height:800px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.svg assert_equals: width: expected 300 but got 1067
FAIL Test video with attribute style=height:800px; and attribute undefined=undefined assert_equals: width: expected 300 but got 1066.65625
PASS Test image with attribute width=500 and attribute height=800 on src /feature-policy/experimental-features/resources/image.jpg
PASS Test image with attribute width=500 and attribute height=800 on src /feature-policy/experimental-features/resources/image.png
PASS Test image with attribute width=500 and attribute height=800 on src /feature-policy/experimental-features/resources/image.svg
PASS Test video with attribute width=500 and attribute height=800
PASS Test image with attribute style=width:500px;height:800px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.jpg
PASS Test image with attribute style=width:500px;height:800px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.png
PASS Test image with attribute style=width:500px;height:800px; and attribute undefined=undefined on src /feature-policy/experimental-features/resources/image.svg
PASS Test video with attribute style=width:500px;height:800px; and attribute undefined=undefined
FAIL Test image size is correctly rendered in iframe of src /feature-policy/experimental-features/resources/feature-policy-image.html assert_equals: width expected 300 but got 64
FAIL Test image size is correctly rendered in iframe of src https://www.web-platform.test:8444/feature-policy/experimental-features/resources/feature-policy-image.html assert_equals: width expected 300 but got 64
Harness: the test ran to completion.
